added search and tags functionality

+++ src/BackEnd/srv.py
@@ -1,49 +1,47 @@
 from flask import Flask, request, abort, make_response, g
-# import mysql.connector, mysql.connector.pooling
-import mysql.connector as mysql
 import json
 import bcrypt
 import uuid
 from flask_cors import CORS
-
-# import Cookies from `js-cookie`
+import mysql.connector, mysql.connector.pooling
 
 # pool = mysql.connector.pooling.MySQLConnectionPool(
-#     # host="my-rds.cmpq1mavlbmq.us-east-1.rds.amazonaws.com",
-#     host='localhost',
-#     user='root',
-#     password='beni2020',
+#     host="my-rds.cmpq1mavlbmq.us-east-1.rds.amazonaws.com",
+#     user='admin',
+#     passwd='beni2020',
 #     database='beni',
 #     buffered=True,
+#     port=3306,
 #     pool_size=3,
+#     pool_name="pool",
 # )
-
-
-db = mysql.connect(
-    host='localhost',
-    user='root',
-    password='Beni',
-    database='beni'
+pool = mysql.connector.pooling.MySQLConnectionPool(
+    host="localhost",
+    user="root",
+    passwd="Beniandsara2020",
+    database="beni",
+    buffered=True,
+    pool_size=3
 )
-
-# print(db)
 
 
 app = Flask(__name__)
 CORS(app)
 
 
-# ,
+# app = Flask(__name__,
 # static_folder='/home/ubuntu/build',
 # static_url_path='/')
-# //to remmember to add db =
 
-# @app.before_request
-# def before_request():
-#     # db = pool.get_connection()
-# @app.teardown_request
-# def teardown_request(exception):
-#     db.close()
+
+@app.before_request
+def before_request():
+    g.db = pool.get_connection()
+
+
+@app.teardown_request
+def teardown_request(exception):
+    g.db.close()
 
 @app.route('/', methods=['GET'])
 def main():
@@ -60,7 +58,7 @@ def login():
     data = request.get_json()
     query = "select id, password from users where name=%s"
     values = (data['username'],)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, values)
     record = cursor.fetchone()
     if not record:
@@ -77,7 +75,7 @@ def login():
     query = "insert into sessions (user_id, session_id) values (%s, %s) on duplicate key update session_id=%s"
     values = (user_id, session_id, session_id)
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
     username = data['username']
     # cookieJs = {"username": username, "userId": user_id}
     # resp = make_response(cookieJs)
@@ -91,10 +89,9 @@ def login():
 @app.route('/signup', methods=['POST'])
 def signup():
     data = request.get_json()
-    print(data)
     query = "select name from users where name = %s"
     values = (data['username'],)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, values)
     record = cursor.fetchone()
     if record:
@@ -104,7 +101,7 @@ def signup():
     hashedPass = bcrypt.hashpw(data['password'].encode('utf-8'), bcrypt.gensalt())
     values = (data['username'], hashedPass, data['useremail'])
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
 
     query = "select id from users where name = %s"
     values = (data['username'],)
@@ -116,24 +113,14 @@ def signup():
     values = (user_id, session_id, session_id)
 
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
     resp = make_response()
     resp.set_cookie("session_id", session_id)
     cursor.close()
     return resp
 
 
-# @app.route('/getidbyname/<name>', methods=['GET'])
-# def get_id(name):
-#     name = str(name)
-#     query = "select id from users where name=%s"
-#     value = (name,)
-#     cursor = db.cursor()
-#     cursor.execute(query, value)
-#     record = cursor.fetchone()
-#     header = ['id']
-#     cursor.close()
-#     return json.dumps(dict(zip(header, record)))
+
 
 @app.route('/logout', methods=['POST'])
 def logout():
@@ -143,9 +130,9 @@ def logout():
     id = data['userId']
     query = "delete from sessions where user_id=%s"
     values = (id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
     cursor.close()
     return "success"
 
@@ -154,7 +141,7 @@ def logout():
 def get_post(id):
     query = "select id, title, content, published, author, imageurl, author_id from posts where id = %s"
     value = (id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, value)
     record = cursor.fetchone()
     if not record:
@@ -168,18 +155,36 @@ def get_post(id):
 @app.route('/posts', methods=['GET'])
 def get_all_posts():
     query = "select id, title, content, published, author, imageurl, author_id from posts"
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query)
     records = cursor.fetchall()
     if not records:
-        cursor.close();
+        cursor.close()
         abort(401)
         # cursor.close()
-
     header = ['id', 'title', 'content', 'published', 'author', 'imageUrl', 'authorId']
     data = []
     for r in records:
         data.append(dict(zip(header, r)))
+    cursor.close()
+    return json.dumps(data)
+
+
+@app.route('/tags/<postId>', methods=['GET'])
+def get_all_tags(postId):
+    query = "select id, name, post_id from tags where post_id=%s"
+    value = (postId,)
+    cursor = g.db.cursor()
+    cursor.execute(query, value)
+    records = cursor.fetchall()
+    if not records:
+        cursor.close()
+        abort(401)
+    header = ['id', 'name', 'post_id']
+    data = []
+    for r in records:
+        data.append(dict(zip(header, r)))
+    cursor.close()
     return json.dumps(data)
 
 
@@ -188,22 +193,34 @@ def add_post():
     data = request.get_json()
     query = "insert into posts (title, content, published, author, imageurl, author_id) values (%s, %s, %s, %s, %s, %s)"
     values = (data['title'], data['content'], data['published'], data['username'], data['imageUrl'], data['userId'])
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
     records = cursor.lastrowid
     if not records:
         abort(401)
         cursor.close()
     add_post_by_id = cursor.lastrowid
     cursor.close()
+
+    tags = data['tags']
+    if not tags:
+        abort(401)
+        cursor.close()
+    for r in tags:
+        queryForTags = "insert into tags (name, post_id) values (%s,%s)"
+        valuesForTags = (r, add_post_by_id)
+        cursor = g.db.cursor()
+        cursor.execute(queryForTags, valuesForTags)
+        g.db.commit()
+        cursor.close()
     return get_post(add_post_by_id)
 
 
 def get_post(id):
     query = "select id, title, content, published, author, imageurl, author_id from posts where id = %s"
     value = (id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, value)
     record = cursor.fetchone()
     if not record:
@@ -214,17 +231,23 @@ def get_post(id):
     return json.dumps(dict(zip(header, record)))
 
 
-@app.route('/edit', methods=['POST'])
-def edit():
+@app.route('/editpost', methods=['POST'])
+def edit_post():
     data = request.get_json()
+    print("data from editpost ==")
+    print(data)
     author_query = "select author_id from posts where id=%s"
     author_value = (data['postId'],)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(author_query, author_value)
     record = cursor.fetchone()
     if not record:
         cursor.close()
         abort(401)
+    print("data['authorId'] ==")
+    print(data['authorId'])
+    print("record[0] ==")
+    print(record[0])
 
     if data['authorId'] != record[0]:
         cursor.close()
@@ -234,8 +257,22 @@ def edit():
     values = (data['title'], data['content'], data['published'], data['author'], data['imageUrl'], data['authorId'],
               data['postId'])
     cursor.execute(query, values)
+    g.db.commit()
     cursor.close()
     return "success edit post "
+
+@app.route('/editcomment', methods=['POST'])
+def edit_comment():
+    data = request.get_json()
+    print("data from edit comment == ")
+    print(data)
+    cursor = g.db.cursor()
+    query = "update comments set title=%s, content=%s, author=%s, author_id=%s, post_id=%s ,published=%s  where id=%s"
+    values = (data['title'], data['content'],  data['author'], data['authorId'], data['postId'], data['published'], data['id'])
+    cursor.execute(query, values)
+    g.db.commit()
+    cursor.close()
+    return "success edit comment "
 
 
 @app.route('/comment/<id>', methods=['GET', 'POST'])
@@ -249,30 +286,32 @@ def manage_requests(id):
 def get_all_comments(id):
     query = "select id, title, content, author, author_id, post_id ,published from comments where post_id=%s"
     value = (id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, value)
     records = cursor.fetchall()
     if not records:
+        cursor.close()
         abort(401)
-        cursor.close();
     header = ['id', 'title', 'content', 'author', 'author_id', 'post_id', 'published']
     data = []
     for r in records:
         data.append(dict(zip(header, r)))
+    cursor.close()
     return json.dumps(data)
 
 
 def add_comment():
     data = request.get_json()
     query = "insert into comments (title, content, author, author_id, post_id,published) values (%s,%s,%s,%s,%s,%s) on duplicate key update post_id = %s"
-    values = (data['title'], data['content'], data['username'], data['authorId'], data['postId'], data['published'], data['postId'])
-    cursor = db.cursor()
+    values = (data['title'], data['content'], data['username'], data['authorId'], data['postId'], data['published'],
+              data['postId'])
+    cursor = g.db.cursor()
     cursor.execute(query, values)
-    db.commit()
+    g.db.commit()
     records = cursor.lastrowid
     if not records:
-        abort(401)
         cursor.close()
+        abort(401)
     add_comment_id = cursor.lastrowid
     cursor.close()
     return get_comment(add_comment_id)
@@ -281,7 +320,7 @@ def add_comment():
 def get_comment(id):
     query = "select title, content, author, published from comments where id = %s"
     value = (id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, value)
     record = cursor.fetchone()
     header = ['title', 'author', 'content', 'published']
@@ -295,18 +334,33 @@ def delete_post_and_comments():
     post_id = data['postId']
     queryComment = "delete from comments where post_id =%s"
     value = (post_id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(queryComment, value)
-    db.commit()
+    if not cursor:
+        cursor.close();
+        abort(401)
+    g.db.commit()
+    cursor.close()
+
+    queryTags = "delete from Tags where post_id =%s"
+    value = (post_id,)
+    cursor = g.db.cursor()
+    cursor.execute(queryTags, value)
+    if not cursor:
+        cursor.close();
+        abort(401)
+    g.db.commit()
     cursor.close()
 
     queryPost = "delete from posts where id = %s"
     value = (post_id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(queryPost, value)
-    db.commit()
+    g.db.commit()
     cursor.close()
     return "Succeeded to delete comments for pt_id"
+
+
 
 @app.route('/deletecomment', methods=['POST'])
 def delete_comments():
@@ -314,32 +368,53 @@ def delete_comments():
     comment_id = data['commentId']
     queryComment = "delete from comments where id =%s"
     value = (comment_id,)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(queryComment, value)
-    db.commit()
+    g.db.commit()
     cursor.close()
     return "Succeeded to delete comment"
 
+
 @app.route('/search/<wordsearch>', methods=['GET'])
 def get_all_posts_for_search(wordsearch):
-    print(wordsearchus
-          )
     query = "select * from posts where content like %s"
     value = ("%" + wordsearch + "%",)
-    cursor = db.cursor()
+    cursor = g.db.cursor()
     cursor.execute(query, value)
     records = cursor.fetchall()
-    print(records)
     if not records:
         cursor.close();
         abort(401)
-        # cursor.close()
 
     header = ['id', 'title', 'content', 'published', 'author', 'imageUrl', 'authorId']
     data = []
     for r in records:
         data.append(dict(zip(header, r)))
+    cursor.close()
     return json.dumps(data)
+
+
+@app.route('/searchtags/<wordsearch>', methods=['GET'])
+def get_all_posts_for_search_tags(wordsearch):
+    query = "select post_id from tags where name like %s"
+    value = ("%" + wordsearch + "%",)
+    cursor = g.db.cursor()
+    cursor.execute(query, value)
+    records = cursor.fetchall()
+    newrecords = list(set(records))
+    if not records:
+        cursor.close()
+        abort(401)
+
+    header = ['id', 'title', 'content', 'published', 'author', 'imageUrl', 'authorId']
+    data = []
+    for r in newrecords:
+        data.append(dict(zip(header, r)))
+    cursor.close()
+    return json.dumps(data)
+
+
+
 
 if __name__ == "__main__":
     app.run()
